TPE_global_optimization/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=py37h27cfd23_1
ase=3.20.1=py_0
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py37heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.30=ha878542_0
cached-property=1.5.2=pypi_0
cairo=1.16.0=hf32fb01_1
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.6=py37h400218f_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cycler=0.10.0=py37_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
flask=1.1.2=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py37_1
geometric=0.9.7.2=pypi_0
glib=2.69.1=h5202010_0
h5py=3.4.0=pypi_0
hyperopt=0.2.5=pyh9f0ad1d_0
icu=58.2=he6710b0_3
importlib-metadata=4.8.1=py37h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.27.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3deb0d_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.0=py37h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h7f8727e_0
jsonpickle=2.0.0=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter-contrib-core=0.3.3=pypi_0
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=1_h6e990d7_netlib
libboost=1.73.0=h3ff78a5_11
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
liblapack=3.9.0=3_h893e4fe_netlib
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py37h27cfd23_0
matplotlib-base=3.4.2=py37hab158f2_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.2=pyhd3eb1b0_0
nglview=3.0.3=pyh8a188c0_0
notebook=6.4.3=py37h06a4308_0
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py37_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.2=py37h8c16a72_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37h2c7a002_0
pip=21.2.2=py37h06a4308_0
pixman=0.40.0=h7f8727e_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-boost=1.73.0=py37ha9443f7_11
pyberny=0.6.3=pypi_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pymongo=3.12.0=py37h295c915_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py37h7b6447c_0
pyscf=1.7.6.post1=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py37h295c915_1
rdkit=2020.09.1.0=py37hd50e099_1
readline=8.1=h27cfd23_0
rmsd=1.4=pyhd8ed1ab_0
scipy=1.7.1=py37h292c36d_2
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
terminado=0.9.4=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
tornado=6.1=py37h27cfd23_0
tqdm=4.57.0=pyhd8ed1ab_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py37_0
xtb=6.3.3=h323e27b_1
xtb-python=20.2=py37h4abf009_2
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
